+++ pauli_tracker/CHANGELOG.md
@@ -21,6 +21,14 @@ methods.
 ### Added
 ### Changed
 ### Deprecated
+### Removed
+### Fixed
+### Security
+
+## [0.4.2] - 2024-09-03
+### Added
+### Changed
+### Deprecated
 - Deprecate the `Pauli::add` method in favour of `Pauli::multiply`.
 ### Removed
 ### Fixed
@@ -213,7 +221,8 @@ them, but hopefully the most important ones.
 
 ## [0.1.0] - 2023-06-01
 
-[Unreleased]: https://github.com/taeruh/pauli_tracker/compare/v0.4.1...HEAD
+[Unreleased]: https://github.com/taeruh/pauli_tracker/compare/v0.4.2...HEAD
+[0.4.2]: https://github.com/taeruh/pauli_tracker/compare/v0.4.1...v0.4.2
 [0.4.1]: https://github.com/taeruh/pauli_tracker/compare/v0.4.0...v0.4.1
 [0.4.0]: https://github.com/taeruh/pauli_tracker/compare/v0.3.2...v0.4.0
 [0.3.2]: https://github.com/taeruh/pauli_tracker/compare/v0.3.1...v0.3.2